optional-requirements.txt
networkx
scipy
matplotlib
rtree
pygraphviz
